Move public connectors arrow methods to their own separate files and colocate tests
