+++ apollo-federation/src/sources/connect/json_selection/methods.rs
@@ -59,7 +59,7 @@ pub(super) enum ArrowMethod {
 macro_rules! impl_arrow_method {
     ($struct_name:ident, $impl_fn_name:ident, $shape_fn_name:ident) => {
         #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-        pub(super) struct $struct_name;
+        pub(crate) struct $struct_name;
         impl $crate::sources::connect::json_selection::methods::ArrowMethodImpl for $struct_name {
             fn apply(
                 &self,
@@ -141,15 +141,15 @@ impl std::ops::Deref for ArrowMethod {
     fn deref(&self) -> &Self::Target {
         match self {
             // Public methods:
-            Self::Echo => &public::EchoMethod,
-            Self::Map => &public::MapMethod,
-            Self::Match => &public::MatchMethod,
-            Self::First => &public::FirstMethod,
-            Self::Last => &public::LastMethod,
-            Self::Slice => &public::SliceMethod,
-            Self::Size => &public::SizeMethod,
-            Self::Entries => &public::EntriesMethod,
-            Self::JsonStringify => &public::JsonStringifyMethod,
+            Self::Echo => &public::echo::EchoMethod,
+            Self::Map => &public::map::MapMethod,
+            Self::Match => &public::r#match::MatchMethod,
+            Self::First => &public::first::FirstMethod,
+            Self::Last => &public::last::LastMethod,
+            Self::Slice => &public::slice::SliceMethod,
+            Self::Size => &public::size::SizeMethod,
+            Self::Entries => &public::entries::EntriesMethod,
+            Self::JsonStringify => &public::json_stringify::JsonStringifyMethod,
 
             // Future methods:
             Self::TypeOf => &future::TypeOfMethod,
