+++ apollo-federation/src/sources/connect/json_selection/methods/public/echo.rs
@@ -1,0 +1,223 @@
+use apollo_compiler::collections::IndexMap;
+use serde_json_bytes::Value as JSON;
+use shape::Shape;
+use shape::location::SourceId;
+
+use crate::impl_arrow_method;
+use crate::sources::connect::json_selection::ApplyToError;
+use crate::sources::connect::json_selection::ApplyToInternal;
+use crate::sources::connect::json_selection::MethodArgs;
+use crate::sources::connect::json_selection::PathList;
+use crate::sources::connect::json_selection::VarsWithPathsMap;
+use crate::sources::connect::json_selection::apply_to::ApplyToResultMethods;
+use crate::sources::connect::json_selection::immutable::InputPath;
+use crate::sources::connect::json_selection::location::Ranged;
+use crate::sources::connect::json_selection::location::WithRange;
+
+impl_arrow_method!(EchoMethod, echo_method, echo_shape);
+/// Echo simply returns back whichever value is provided in it's arg.
+/// The simplest possible case is $.echo("hello world") which would result in "hello world"
+///
+/// However, it will also reflect back any type passed into it allowing you to act on those:
+///
+/// $->echo([1,2,3])->first         would result in "1"
+///
+/// It's also worth noting that you can use $ to refer to to the selection and pass that into echo and you can also use @ to refer to the value that echo is being run on.
+///
+/// For example, assuming my selection is { firstName: "John", children: ["Jack"] }...
+///
+/// $->echo($.firstName)            would result in "John"
+/// $.children->echo(@->first)      would result in "Jack"
+fn echo_method(
+    method_name: &WithRange<String>,
+    method_args: Option<&MethodArgs>,
+    data: &JSON,
+    vars: &VarsWithPathsMap,
+    input_path: &InputPath<JSON>,
+    tail: &WithRange<PathList>,
+) -> (Option<JSON>, Vec<ApplyToError>) {
+    if let Some(MethodArgs { args, .. }) = method_args {
+        if let Some(arg) = args.first() {
+            return arg
+                .apply_to_path(data, vars, input_path)
+                .and_then_collecting_errors(|value| tail.apply_to_path(value, vars, input_path));
+        }
+    }
+    (
+        None,
+        vec![ApplyToError::new(
+            format!("Method ->{} requires one argument", method_name.as_ref()),
+            input_path.to_vec(),
+            method_name.range(),
+        )],
+    )
+}
+#[allow(dead_code)] // method type-checking disabled until we add name resolution
+fn echo_shape(
+    method_name: &WithRange<String>,
+    method_args: Option<&MethodArgs>,
+    input_shape: Shape,
+    dollar_shape: Shape,
+    named_var_shapes: &IndexMap<&str, Shape>,
+    source_id: &SourceId,
+) -> Shape {
+    if let Some(first_arg) = method_args.and_then(|args| args.args.first()) {
+        return first_arg.compute_output_shape(
+            input_shape,
+            dollar_shape,
+            named_var_shapes,
+            source_id,
+        );
+    }
+    Shape::error(
+        format!("Method ->{} requires one argument", method_name.as_ref()),
+        method_name.shape_location(source_id),
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json_bytes::json;
+
+    use crate::selection;
+
+    #[test]
+    fn echo_should_output_value_when_applied_to_null() {
+        assert_eq!(
+            selection!("$->echo('oyez')").apply_to(&json!(null)),
+            (Some(json!("oyez")), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_output_value_when_applied_to_array() {
+        assert_eq!(
+            selection!("$->echo('oyez')").apply_to(&json!([1, 2, 3])),
+            (Some(json!("oyez")), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_allow_selection_from_array_value() {
+        assert_eq!(
+            selection!("$->echo([1, 2, 3]) { id: $ }").apply_to(&json!(null)),
+            (Some(json!([{ "id": 1 }, { "id": 2 }, { "id": 3 }])), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_allow_arrow_methods_off_returned_value() {
+        assert_eq!(
+            selection!("$->echo([1, 2, 3])->last { id: $ }").apply_to(&json!(null)),
+            (Some(json!({ "id": 3 })), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_allow_at_sign_to_input_value_from_selection_function() {
+        assert_eq!(
+            selection!("$.nested.value->echo(['before', @, 'after'])").apply_to(&json!({
+                "nested": {
+                    "value": 123,
+                },
+            })),
+            (Some(json!(["before", 123, "after"])), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_allow_dollar_sign_to_input_applied_to_value() {
+        assert_eq!(
+            selection!("$.nested.value->echo(['before', $, 'after'])").apply_to(&json!({
+                "nested": {
+                    "value": 123,
+                },
+            })),
+            (
+                Some(json!(["before", {
+            "nested": {
+                "value": 123,
+            },
+        }, "after"])),
+                vec![]
+            ),
+        );
+    }
+
+    #[test]
+    fn echo_should_allow_selection_functions_result_passed_as_value() {
+        assert_eq!(
+            selection!("results->echo(@->first)").apply_to(&json!({
+                "results": [
+                    [1, 2, 3],
+                    "ignored",
+                ],
+            })),
+            (Some(json!([1, 2, 3])), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_allow_arrow_functions_on_result_of_echo() {
+        assert_eq!(
+            selection!("results->echo(@->first)->last").apply_to(&json!({
+                "results": [
+                    [1, 2, 3],
+                    "ignored",
+                ],
+            })),
+            (Some(json!(3)), vec![]),
+        );
+    }
+
+    #[test]
+    fn echo_should_not_error_with_trailing_commas() {
+        let nested_value_data = json!({
+            "nested": {
+                "value": 123,
+            },
+        });
+
+        let expected = (Some(json!({ "wrapped": 123 })), vec![]);
+
+        let check = |selection: &str| {
+            assert_eq!(selection!(selection).apply_to(&nested_value_data), expected,);
+        };
+
+        check("nested.value->echo({ wrapped: @ })");
+        check("nested.value->echo({ wrapped: @,})");
+        check("nested.value->echo({ wrapped: @,},)");
+        check("nested.value->echo({ wrapped: @},)");
+        check("nested.value->echo({ wrapped: @ , } , )");
+    }
+
+    #[test]
+    fn echo_should_flatted_object_list_using_at_sign() {
+        // Turn a list of { name, hobby } objects into a single { names: [...],
+        // hobbies: [...] } object.
+        assert_eq!(
+            selection!(
+                r#"
+            people->echo({
+                names: @.name,
+                hobbies: @.hobby,
+            })
+            "#
+            )
+            .apply_to(&json!({
+                "people": [
+                    { "name": "Alice", "hobby": "reading" },
+                    { "name": "Bob", "hobby": "fishing" },
+                    { "hobby": "painting", "name": "Charlie" },
+                ],
+            })),
+            (
+                Some(json!({
+                    "names": ["Alice", "Bob", "Charlie"],
+                    "hobbies": ["reading", "fishing", "painting"],
+                })),
+                vec![],
+            ),
+        );
+    }
+}
