+++ apollo-federation/src/sources/connect/json_selection/methods/public/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod echo;
+pub(crate) mod entries;
+pub(crate) mod first;
+pub(crate) mod json_stringify;
+pub(crate) mod last;
+pub(crate) mod map;
+pub(crate) mod r#match;
+pub(crate) mod size;
+pub(crate) mod slice;
