+++ apollo-federation/src/sources/connect/json_selection/methods/tests.rs
@@ -5,138 +5,6 @@ use super::*;
 use crate::selection;
 
 #[test]
-fn test_echo_method() {
-    assert_eq!(
-        selection!("$->echo('oyez')").apply_to(&json!(null)),
-        (Some(json!("oyez")), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->echo('oyez')").apply_to(&json!([1, 2, 3])),
-        (Some(json!("oyez")), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->echo([1, 2, 3]) { id: $ }").apply_to(&json!(null)),
-        (Some(json!([{ "id": 1 }, { "id": 2 }, { "id": 3 }])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->echo([1, 2, 3])->last { id: $ }").apply_to(&json!(null)),
-        (Some(json!({ "id": 3 })), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->echo([1.1, 0.2, -3.3]) { id: $ }").apply_to(&json!(null)),
-        (
-            Some(json!([{ "id": 1.1 }, { "id": 0.2 }, { "id": -3.3 }])),
-            vec![]
-        ),
-    );
-
-    assert_eq!(
-        selection!("$.nested.value->echo(['before', @, 'after'])").apply_to(&json!({
-            "nested": {
-                "value": 123,
-            },
-        })),
-        (Some(json!(["before", 123, "after"])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$.nested.value->echo(['before', $, 'after'])").apply_to(&json!({
-            "nested": {
-                "value": 123,
-            },
-        })),
-        (
-            Some(json!(["before", {
-            "nested": {
-                "value": 123,
-            },
-        }, "after"])),
-            vec![]
-        ),
-    );
-
-    assert_eq!(
-        selection!("data->echo(@.results->last)").apply_to(&json!({
-            "data": {
-                "results": [1, 2, 3],
-            },
-        })),
-        (Some(json!(3)), vec![]),
-    );
-
-    assert_eq!(
-        selection!("results->echo(@->first)").apply_to(&json!({
-            "results": [
-                [1, 2, 3],
-                "ignored",
-            ],
-        })),
-        (Some(json!([1, 2, 3])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("results->echo(@->first)->last").apply_to(&json!({
-            "results": [
-                [1, 2, 3],
-                "ignored",
-            ],
-        })),
-        (Some(json!(3)), vec![]),
-    );
-
-    {
-        let nested_value_data = json!({
-            "nested": {
-                "value": 123,
-            },
-        });
-
-        let expected = (Some(json!({ "wrapped": 123 })), vec![]);
-
-        let check = |selection: &str| {
-            assert_eq!(selection!(selection).apply_to(&nested_value_data), expected,);
-        };
-
-        check("nested.value->echo({ wrapped: @ })");
-        check("nested.value->echo({ wrapped: @,})");
-        check("nested.value->echo({ wrapped: @,},)");
-        check("nested.value->echo({ wrapped: @},)");
-        check("nested.value->echo({ wrapped: @ , } , )");
-    }
-
-    // Turn a list of { name, hobby } objects into a single { names: [...],
-    // hobbies: [...] } object.
-    assert_eq!(
-        selection!(
-            r#"
-            people->echo({
-                names: @.name,
-                hobbies: @.hobby,
-            })
-            "#
-        )
-        .apply_to(&json!({
-            "people": [
-                { "name": "Alice", "hobby": "reading" },
-                { "name": "Bob", "hobby": "fishing" },
-                { "hobby": "painting", "name": "Charlie" },
-            ],
-        })),
-        (
-            Some(json!({
-                "names": ["Alice", "Bob", "Charlie"],
-                "hobbies": ["reading", "fishing", "painting"],
-            })),
-            vec![],
-        ),
-    );
-}
-
-#[test]
 fn test_typeof_method() {
     fn check(selection: &str, data: &JSON, expected_type: &str) {
         assert_eq!(
@@ -153,72 +21,6 @@ fn test_typeof_method() {
     check("$->typeof", &json!("hello"), "string");
     check("$->typeof", &json!([1, 2, 3]), "array");
     check("$->typeof", &json!({ "key": "value" }), "object");
-}
-
-#[test]
-fn test_map_method() {
-    assert_eq!(
-        selection!("$->map(@->add(10))").apply_to(&json!([1, 2, 3])),
-        (Some(json!(vec![11, 12, 13])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("messages->map(@.role)").apply_to(&json!({
-            "messages": [
-                { "role": "admin" },
-                { "role": "user" },
-                { "role": "guest" },
-            ],
-        })),
-        (Some(json!(["admin", "user", "guest"])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("messages->map(@.roles)").apply_to(&json!({
-            "messages": [
-                { "roles": ["admin"] },
-                { "roles": ["user", "guest"] },
-            ],
-        })),
-        (Some(json!([["admin"], ["user", "guest"]])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("values->map(@->typeof)").apply_to(&json!({
-            "values": [1, 2.5, "hello", true, null, [], {}],
-        })),
-        (
-            Some(json!([
-                "number", "number", "string", "boolean", "null", "array", "object"
-            ])),
-            vec![],
-        ),
-    );
-
-    assert_eq!(
-        selection!("singleValue->map(@->mul(10))").apply_to(&json!({
-            "singleValue": 123,
-        })),
-        (Some(json!([1230])), vec![]),
-    );
-
-    //  TODO: re-test once method type checking is re-enabled
-    // {
-    //     let single_value_data = json!({
-    //         "singleValue": 123,
-    //     });
-    //     let json_selection = selection!("singleValue->map(@->jsonStringify)");
-    //     assert_eq!(
-    //         json_selection.apply_to(&single_value_data),
-    //         (Some(json!(["123"])), vec![]),
-    //     );
-    //     let output_shape = json_selection.compute_output_shape(
-    //         Shape::from_json_bytes(&single_value_data),
-    //         &IndexMap::default(),
-    //         &SourceId::new("test"),
-    //     );
-    //     assert_eq!(output_shape.pretty_print(), "List<String>");
-    // }
 }
 
 #[test]
@@ -241,78 +43,7 @@ fn test_missing_method() {
 }
 
 #[test]
-fn test_match_methods() {
-    assert_eq!(
-        selection!(
-            r#"
-            name
-            __typename: kind->match(
-                ['dog', 'Canine'],
-                ['cat', 'Feline']
-            )
-            "#
-        )
-        .apply_to(&json!({
-            "kind": "cat",
-            "name": "Whiskers",
-        })),
-        (
-            Some(json!({
-                "__typename": "Feline",
-                "name": "Whiskers",
-            })),
-            vec![],
-        ),
-    );
-
-    assert_eq!(
-        selection!(
-            r#"
-            name
-            __typename: kind->match(
-                ['dog', 'Canine'],
-                ['cat', 'Feline'],
-                [@, 'Exotic'],
-            )
-            "#
-        )
-        .apply_to(&json!({
-            "kind": "axlotl",
-            "name": "Gulpy",
-        })),
-        (
-            Some(json!({
-                "__typename": "Exotic",
-                "name": "Gulpy",
-            })),
-            vec![],
-        ),
-    );
-
-    assert_eq!(
-        selection!(
-            r#"
-            name
-            __typename: kind->match(
-                ['dog', 'Canine'],
-                ['cat', 'Feline'],
-                [@, 'Exotic'],
-            )
-            "#
-        )
-        .apply_to(&json!({
-            "kind": "dog",
-            "name": "Laika",
-        })),
-        (
-            Some(json!({
-                "__typename": "Canine",
-                "name": "Laika",
-            })),
-            vec![],
-        ),
-    );
-
+fn test_match_if_methods() {
     assert_eq!(
         selection!(
             r#"
@@ -487,20 +218,6 @@ fn test_arithmetic_methods() {
 #[test]
 fn test_array_methods() {
     assert_eq!(
-        selection!("$->first").apply_to(&json!([1, 2, 3])),
-        (Some(json!(1)), vec![]),
-    );
-
-    assert_eq!(selection!("$->first").apply_to(&json!([])), (None, vec![]),);
-
-    assert_eq!(
-        selection!("$->last").apply_to(&json!([1, 2, 3])),
-        (Some(json!(3)), vec![]),
-    );
-
-    assert_eq!(selection!("$->last").apply_to(&json!([])), (None, vec![]),);
-
-    assert_eq!(
         selection!("$->get(1)").apply_to(&json!([1, 2, 3])),
         (Some(json!(2)), vec![]),
     );
@@ -576,77 +293,6 @@ fn test_array_methods() {
     assert_eq!(
         selection!("$->has(5)").apply_to(&json!([1, 2, 3])),
         (Some(json!(false)), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->slice(1, 3)").apply_to(&json!([1, 2, 3, 4, 5])),
-        (Some(json!([2, 3])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->slice(1, 3)").apply_to(&json!([1, 2])),
-        (Some(json!([2])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->slice(1, 3)").apply_to(&json!([1])),
-        (Some(json!([])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->slice(1, 3)").apply_to(&json!([])),
-        (Some(json!([])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->size").apply_to(&json!([])),
-        (Some(json!(0)), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->size").apply_to(&json!([1, 2, 3])),
-        (Some(json!(3)), vec![]),
-    );
-}
-
-#[test]
-fn test_size_method_errors() {
-    assert_eq!(
-        selection!("$->size").apply_to(&json!(null)),
-        (
-            None,
-            vec![ApplyToError::from_json(&json!({
-                "message": "Method ->size requires an array, string, or object input, not null",
-                "path": ["->size"],
-                "range": [3, 7],
-            }))]
-        ),
-    );
-
-    assert_eq!(
-        selection!("$->size").apply_to(&json!(true)),
-        (
-            None,
-            vec![ApplyToError::from_json(&json!({
-                "message": "Method ->size requires an array, string, or object input, not boolean",
-                "path": ["->size"],
-                "range": [3, 7],
-            }))]
-        ),
-    );
-
-    assert_eq!(
-        selection!("count->size").apply_to(&json!({
-            "count": 123,
-        })),
-        (
-            None,
-            vec![ApplyToError::from_json(&json!({
-                "message": "Method ->size requires an array, string, or object input, not number",
-                "path": ["count", "->size"],
-                "range": [7, 11],
-            }))]
-        ),
     );
 }
 
@@ -987,65 +633,6 @@ fn test_object_methods() {
             }))]
         ),
     );
-
-    assert_eq!(
-        selection!("$->entries").apply_to(&json!({
-            "a": 1,
-            "b": "two",
-            "c": false,
-        })),
-        (
-            Some(json!([
-                { "key": "a", "value": 1 },
-                { "key": "b", "value": "two" },
-                { "key": "c", "value": false },
-            ])),
-            vec![],
-        ),
-    );
-
-    assert_eq!(
-        // This is just like $->keys, given the automatic array mapping of
-        // .key, though you probably want to use ->keys directly because it
-        // avoids cloning all the values unnecessarily.
-        selection!("$->entries.key").apply_to(&json!({
-            "one": 1,
-            "two": 2,
-            "three": 3,
-        })),
-        (Some(json!(["one", "two", "three"])), vec![]),
-    );
-
-    assert_eq!(
-        // This is just like $->values, given the automatic array mapping of
-        // .value, though you probably want to use ->values directly because
-        // it avoids cloning all the keys unnecessarily.
-        selection!("$->entries.value").apply_to(&json!({
-            "one": 1,
-            "two": 2,
-            "three": 3,
-        })),
-        (Some(json!([1, 2, 3])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("$->entries").apply_to(&json!({})),
-        (Some(json!([])), vec![]),
-    );
-
-    assert_eq!(
-        selection!("notAnObject->entries").apply_to(&json!({
-            "notAnObject": true,
-        })),
-        (
-            None,
-            vec![ApplyToError::from_json(&json!({
-                "message": "Method ->entries requires an object input, not boolean",
-                "path": ["notAnObject", "->entries"],
-                "range": [13, 20],
-            }))]
-        ),
-    );
 }
 
 #[test]
@@ -1154,41 +741,5 @@ fn test_logical_methods() {
             },
         })),
         (Some(json!(false)), vec![]),
-    );
-}
-
-#[rstest::rstest]
-#[case(json!(null), json!("null"), vec![])]
-#[case(json!(true), json!("true"), vec![])]
-#[case(json!(false), json!("false"), vec![])]
-#[case(json!(42), json!("42"), vec![])]
-#[case(json!(10.8), json!("10.8"), vec![])]
-#[case(json!("hello world"), json!("\"hello world\""), vec![])]
-#[case(json!([1, 2, 3]), json!("[1,2,3]"), vec![])]
-#[case(json!({ "key": "value" }), json!("{\"key\":\"value\"}"), vec![])]
-#[case(json!([1, "two", true, null]), json!("[1,\"two\",true,null]"), vec![])]
-fn table_test_json_stringify_method(
-    #[case] input: JSON,
-    #[case] expected: JSON,
-    #[case] errors: Vec<ApplyToError>,
-) {
-    assert_eq!(
-        selection!("$->jsonStringify").apply_to(&input),
-        (Some(expected), errors),
-    );
-}
-
-#[test]
-fn test_json_stringify_method_error() {
-    assert_eq!(
-        selection!("$->jsonStringify(1)").apply_to(&json!(null)),
-        (
-            None,
-            vec![ApplyToError::new(
-                "Method ->jsonStringify does not take any arguments".to_string(),
-                vec![json!("->jsonStringify")],
-                Some(3..16)
-            )]
-        ),
     );
 }
